// src/types/database.types.ts
export type Database = {
  public: {
    Tables: {
// ...
      stock_movements: {
        Row: {
          created_at: string
          id: string
          quantity: number
          reason: string | null
          shop_id: string
          type: "IN" | "OUT" | "WASTE" | "RETURN"
          variant_id: string
        }
        Insert: {
          created_at?: string
          id?: string
          quantity: number
          reason?: string | null
          shop_id: string
          type: "IN" | "OUT" | "WASTE" | "RETURN"
          variant_id: string
        }
        Update: {
          created_at?: string
          id?: string
          quantity?: number
          reason?: string | null
          shop_id?: string
          type?: "IN" | "OUT" | "WASTE" | "RETURN"
          variant_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "stock_movements_shop_id_fkey"
            columns: ["shop_id"]
            isOneToOne: false
            referencedRelation: "shops"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "stock_movements_variant_id_fkey"
            columns: ["variant_id"]
            isOneToOne: false
            referencedRelation: "product_variants"
            referencedColumns: ["id"]
          }
        ]
      }
    }
// ...
  }
}